+++ src/main/java/com/github/games647/flexiblelogin/listener/PreventListener.java
@@ -106,7 +106,6 @@ public class PreventListener {
 
     @Listener
     public void onPlayerDamage(DamageEntityEvent damageEntityEvent, @First Player player) {
-        //checks the cause
         checkLoginStatus(damageEntityEvent, player);
     }