+++ src/main/java/com/github/games647/flexiblelogin/tasks/LoginTask.java
@@ -59,6 +59,11 @@ public class LoginTask implements Runnable {
                 if (plugin.getConfigManager().getConfig().isUpdateLoginStatus()) {
                     plugin.getDatabase().flushLoginStatus(account, true);
                 }
+                
+                plugin.getGame().getScheduler().createTaskBuilder()
+                        .async()
+                        .execute(() -> plugin.getProtectionManager().unprotect(player))
+                        .submit(plugin);
             } else {
                 player.sendMessage(plugin.getConfigManager().getConfig().getTextConfig().getIncorrectPassword());
             }