+++ src/main/java/com/github/games647/flexiblelogin/tasks/RegisterTask.java
@@ -63,6 +63,11 @@ public class RegisterTask implements Runnable {
                 if (plugin.getConfigManager().getConfig().isUpdateLoginStatus()) {
                     plugin.getDatabase().flushLoginStatus(createdAccount, true);
                 }
+
+                plugin.getGame().getScheduler().createTaskBuilder()
+                        .async()
+                        .execute(() -> plugin.getProtectionManager().unprotect(player))
+                        .submit(plugin);
             } catch (Exception ex) {
                 plugin.getLogger().error("Error creating hash", ex);
                 player.sendMessage(plugin.getConfigManager().getConfig().getTextConfig().getErrorCommandMessage());